Replace some uses of `Array` by `Vector` or `Matrix` (#60154)
Progress towards issue #56922 by @LilithHafner. Turns out I didn't find
many uses of `Array{T, 1}` or `Array{T, 2}` in `doc` or `base` that were
"clearly" in need of being updated to `Vector` or `Matrix`.

There are a bunch that can't be changed because they occur before
`Vector` or `Matrix` are even defined.

There are a couple borderline cases, and I may have overlooked some, but
most I looked at felt like "I could change it to Vector or Matrix but I
think there is a didactical point in using `Array` here. E.g. in the
definition of `struct BitVector <: AbstractArray{Bool, 1}; end` I think
the context makes it desirable to leave it at that instead of changing
it to `Vector` -- but I admit this is subjective.

There was more in `test`, though that's perhaps less important.

All in all I think we could close #56922