[JuliaLowering] Remove use of `@SyntaxTree` in ccall (#60192)
I'm assuming this macro (which is defined as "go find the file or REPL history
entry this Expr came from and re-parse the text to SyntaxTree") is mostly
for convenience in interactive environments. It's also breakable at
precompile time with changes in the way we expand `quote`, and it takes
a noticable amount of time in inference.  This change just uses the
less-convenient `@ast` form like the rest of the project.