Use jlutilities mechanism for external package required in test
Otherwise we try to install the Pkg into the stdlib depot, which
corrupts state for Pkg tests. If the test is run manually from
`Base.runtests` in a deployed setting, it'll be skipped unless
manually installed by the user. This matches how Pkg.jl handles
this. After this merges, we should tweak CI to pre-install the
packages for both this and the Pkg.jl tests.

Fixes intermittent failures in https://github.com/JuliaLang/julia/pull/60222