Don't introduce a block around `y` in parsing `x' = y` (#60246)
Closes #59911. `'` is generally treated as its own thing in Expr and
flisp, but treated as a special type of call in JuliaSyntax. This change
makes `x' = y` parse to the <=1.11 `(= (|'| x) y)` instead of wrapping
the body in a block like we do when parsing `f(x) = y`.

The special case is fine to me given that it's a more understandable
parse, and we're hoping to move away from LineNumberNodes in the future
anyway. (I think this was the original reason for the block-wrap for
`f(x) = y`).