Avoid using `fork()` when probing system `libstdc++`
Rather than relying on the linker to do all of the heavy lifting, this
changes our probing technique to directly parse the `ld.so.cache` and
`libstdc++.so.6` files. As long as we can expect `ld.so.cache` to be
available in `/etc` on all systems, this seems to be a reliable way to
locate system libraries on Linux systems.

That allows us to avoid `fork()`. For a small application that has just
started up `fork()` is no big deal, but it's quite heavy-handed for Julia-
as-a-library scenarios where resident memory may already be large.

Many soft-embedded targets also do not support fork() well, so it is
good for our compatibility to adjust this.