Define inference lattice types in boot.jl rather than jltypes
As a general rule, we should only define things in jltypes
that cannot be expressed in ordinary julia code. These inference
lattice types do not fall into that category. They need to be in
core (at the moment) because codegen looks at them, but there is
no reason to define them in jltypes.c.