loading: always append bundled depots to DEPOT_PATH
When JULIA_DEPOT_PATH is set, the bundled depots are now always appended
to DEPOT_PATH, removing the common footgun of forgetting the trailing
':' or ';'. A new environment variable JULIA_DEPOT_PATH_BUNDLED can be
set to false to disable this behavior for users who explicitly want to
exclude bundled resources.