Handle large read/write/sendfile to File; return actual bytes written
The return value for the `jl_fs_write/read/sendfile` functions is an `int`, even
through they all take `size_t` length arguments, resulting in potentially
truncated counts for large file operations.  Instead, on success (including
partial read/write), we should return the `uv_fs_t.result` field.

I have added a test for this but have left it disabled because it allocates a 5
GiB Vector to read into.

One consequence of the truncated return values is that we previously mangled
large files while copying them with sendfile.  If the return value is truncated
to a random negative value, it manifests as #30723 or #39868.  If it is
truncated to a positive value, sendfile will create a very large destination
file with thousands of repeated sections of the source file, causing #56537.
The bug will only manifest if LLVM decides to zero-extend the return value.

This also changes `jl_fs_write` to return the actual number of bytes written,
rather than the number of bytes it was asked to write.  If a write returns
EAGAIN (usually because the file is O_NONBLOCK/O_NODELAY), we never see the
actual number of bytes written.