Don't create a subprocess in the REPL precompile script
We use raw Base.Filesystem.File IO on PTYs in the REPL precompile script.  Due
to #24440, spawning a subprocess using the REPL's shell mode clears the
O_NONBLOCK flag set on the PTY, resulting in deadlocks when libuv later reads or
writes to it from Julia code.  I haven't identified what caused this to become
more of a problem recently, but I don't think some extra compilation latency
when starting shell mode is worth it if it's causing mystery timeouts in CI.