added a few assertions in the GC to document a lock-free list invariant in code (#60324)
See PR title. I believe it's a good idea to document this property in
code with an assertion, since it's a tricky invariant that makes it safe
to use our current lock-free list implementation.

In the long term, it might be better to use a lock-free list with an
actual ABA-prevention mechanism to make the code more robust.