RFC: syntax: Implement for/while-then & labeled, multi-level break
This implements multi-level and labeled break as contemplated in #5334.
In addition this adds support for break-with-value (#22891) as well
as for-then (aka for-else #1289). Also while-then of course.
All three features are syntax gated to 1.14 syntax.

The syntax for multi-level break is as follows:
```
for i = 1:10
  for j = 1:10
    break break (i, j)
  end
end # Loop evaluate to `(1,1)
```

The break value can be `continue` in which case the next innermost
loop is continued:

```
julia> for i = 1:3
         for j = 1:3
           i > 1 && j == 2 && break continue
           @show (i,j)
         end
         @show i
       end
(i, j) = (1, 1)
(i, j) = (1, 2)
(i, j) = (1, 3)
i = 1
(i, j) = (2, 1)
(i, j) = (3, 1)
```

For more deeply nested loops, the loop can be annotated with a
`@label` and the the break or continue can be targeted using `@goto`:

```
julia> @label outer for a = 1:2
         for b = 1:3
           for c = 1:4
             b > 1 && c == 2 && @goto outer break
           end
           @show b
         end
         @show a
      end
b = 1
```

Naturally `continue` is supported here as well.

The syntax and semantics for `for-then` are as proposed in the
issue:
```
function has5(iter)
  return for x in iter
    x == 5 && break true
  then
    false
  end
end
```

Any `break` (including multi-level and labeled) skips the corresponding
loop's `then` block, which ordinarily would run at loop completion.