Set max_methods 1 for several IO functions
This reduces the number of invalidations of loading BufferIO.jl commit 957f141
from 1026 to 549. The large number of invalidations come from excessive world
splitting on un-inferred IO objects, which, ironically, are probably
intentionally left un-inferred to reduce unnecessary compilation.

This is a poor bandaid solution: World splitting delenda est and all that.
If that is not feasible, then an option to selectively disable it e.g. through
a future `@max_methods 0` would impove the situation. For example, about 400
methods are invalidated due to the addition of a new `write(::T, ::Char)`
method, which invalidates the single, world split `write(::IO, ::Char)` method.

However, deeper fixes of this issue is outside the scope of this PR.