`Expr`/`SyntaxTree` parity: trivial conversions (#60373)
As discussed on slack with @c42f and @topolarity, I gave conversions an
honest try, but couldn't see the end of the tunnel in terms of
correctness. The equivalence between SyntaxNode and Expr is only defined
on surface syntax or fully-macro-expanded syntax, and not the steps in
between, or if errors occur. We have a plan for syntax evolution now,
too, so there compelling reasons not to lump changes to AST layout into
JuliaLowering's macro expansion.