Avoid `Attribute::StackProtectStrong` for AArch64/RISC-V debug builds
When using a code model other than large, e.g. on AArch64 (small) or
RISC-V (medium), the stack canary can end up relocated outside of the
range permissible by the code model, as observed in issue 59303. To fix
this, we could either use a large code model everywhere or disable
strong stack protection for debug builds on AArch64 and RISC-V. The
former is not ideal as some combinations of platform and code model are
not as well-maintained upstream in LLVM, so let's do the latter.

Fixes issue 59303

Co-Authored-By: Sam Schweigel <sam.schweigel@juliahub.com>