Fix JET warning in `_spawn_primitive` (#60452)
In various packages I am JETing, I see the same error being reported:

> local variable `pp` may be undefined: pp::Base.Process

This patch helps JET proof that there is no actual problem.

Would be kinda nice to see this backported to 1.12 so that if I run JET
with 1.12.4 this is not reported anymore. And of course also to 1.13.