[Windows] Use `TimerQueue` API to register Profile watchdog (#60463)
On x86-64 Windows, we're unfortunately pretty reliant on this timer to
avoid (otherwise unavoidable) deadlocks between `RtlLookupFunctionEntry`
and many other internal functions in the Windows runtime (incl.
`RtlAllocateHeap` and `LdrLoadDll`).

For unknown reasons, `RegisterWaitForSingleObject` seems to be quite bad
about dropping the callback on the floor in contended situations (we
deadlock but the timer callback is simply never called). This
alternative API appears to be better-behaved.

This does not solve the deadlock identified by @xal-0 in
https://github.com/JuliaLang/julia/pull/60454#issuecomment-3687877148
(or several others)