Don't make graph type abstract in SyntaxTree (#60430)
I didn't notice we had both (1) abstract attribute storage in
SyntaxGraph and (2) abstract graph type in SyntaxTree and other structs.
This PR keeps the first (it's established that we might experiment with this
type to try and improve performance), but removes the second, as there's
only one graph type that works with SyntaxTree (which is just a graph and
integer node ID). It would be easier to define a new tree type than support
a different graph in SyntaxTree.

Noticed while scrolling through all methodinstances and seeing a weird
mix of partially-parameterized SyntaxTrees in signatures; reduces precompilation
time by about 25% (we are down to less than double JuliaSyntax now!)