Compiler frontend API / lowering iterator
Here I introduce a new compiler frontend API aiming to make the use of
`Expr` optional when implementing `include()`. This paves the way for
preserving full expression provenance using JuliaSyntax+JuliaLowering in
the future making it possible to replace `Expr` with `SyntaxTree`.

To make this possible, we need to eliminate the runtime's reliance on
*interpreting* certain `Expr` forms, in particular `toplevel` and
`module`. Instead, the iteration over toplevel/module expressions is
factored into a frontend API with the `AbstractCompilerFrontend` and
`TopLevelCodeIterator` interfaces.

A `simple_eval()` function is defined, allowing the runtime to implement
evaluation as a sequence of
* Thunk evaluations with CodeInfo (LoweredValue used instead of CodeInfo
  for trivial cases though we perhaps should deprecate that)
* BeginModule evaluations which open a new module statement
* EndModule evaluations which close a module statement

Having separated expression iteration from evaluation it's also possible
to implement some forms of nonstandard evaluation with this interface.

Included here is an implementation of the frontend which uses the flisp
implementation of parsing and lowering.

TODO: Create a frontend which mirrors the current default (JuliaSyntax
for parsing, flisp for lowering) and a frontend which uses
JuliaSyntax+JuliaLowering to avoid ever creating Expr unless necessary.