Compiler frontend API / lowering iterator
Here we introduce a new compiler frontend API aiming to make the use of
`Expr` optional when implementing `include()`. This paves the way for
preserving full expression provenance using JuliaSyntax+JuliaLowering in
the future by making it possible to replace `Expr` with `SyntaxTree`.

To make this possible we need to eliminate the runtime's reliance on
`Expr` when interpreting `toplevel` and `module` forms. Instead, the
iteration over toplevel/module expressions is factored into a frontend
with the `TopLevelCodeIterator` interface.

We also want some level of integration between parsing and lowering,
allowing us to avoid conversion to Expr as a common intermediate data
structure. For this, introduce an `AbstractCompilerFrontend` API,
implementations of which can use their own data expression data
structure inside include_string().

`eval()` is implemented by iterating over `TopLevelCodeIterator` and
evaluating each element of the iteration in the latest world. We iterate
over both top level and module bodies, calling `simple_eval()` on a
sequence of
* `CodeInfo` for top level thunks
* `BeginModule` to open a new module
* `EndModule` to close a module
* `ToplevelExpression` to accommodate "top level only" expressions
  (These could - and probably should - be `CodeInfo`, but flisp lowering
  has some remaining special cases which are only represented as Expr.
  JuliaLowering will probably not use ToplevelExprssion)

Included here are two implementations of AbstractCompilerFrontend:
1. One which uses only flisp for parsing + lowering and can be used
   earlier in bootstrap.
2. One which uses JuliaSyntax for parsing and flisp for lowering (the
   current default frontend behavior)

These tools are used to implement all of the frontend-related language
infrastructure and APIs:

* `Base.include_string()` directly calls the frontend include_string()
* `Base.eval()` directly calls the frontend eval()
* `Meta.parse()` and `Meta.lower()` call the frontend directly

More internally,
* `Core.eval()` is implemented using a new hook. Quite a lot of
  packages in the ecosystem call `Core.eval()` rather than the
  two-argument `Base.eval()` so it seems we need to support this to
  allow the bootstrap implementation of Core.eval() to be swapped out.
* `jl_parse` is implemented with a slight modification of the existing
  `Core._parse` hook

Syntax versioning is reimplemented in terms of CompilerFrontend:

Where a module has syntax version differing from the current VERSION, it
gets a compiler frontend instance in the special hidden variable
`var"#compiler-frontend#"` which replaces the `_internal_julia_parser` and
`_internal_julia_lowerer` hooks (TODO: some elements of the latter still
exist in the code)

The Core._parse hook is slightly modified to take a `versionctx` as the
last argument - this may be a module allowing the correct frontend to be
looked up. Parsing functions in Meta also grow a `versionctx` keyword
argument to replace `mod` (versionctx may be a `VersionNumber` to
specify a particular syntax version)